dhcp: simplify condition checks in dhcpX_fail
The conditions checked in dhcpX_fail have been inverted: this allowed
to drop the check on "num_tries_left" used to see if we already had
an expire event (priv->dhcp4.num_tries_left == DHCP_NUM_TRIES_MAX).
The underling assumption that allows this change is that when we get a
timeout dhcp event and we are in "ipx_state" == IP_DONE we are coming
from a failure to renew a lease.